Fix integer division in Resolution.getAspectRatio

+++ src/main/java/org/osm2world/core/util/Resolution.java
@@ -49,7 +49,7 @@ public class Resolution {
 	}
 
 	public float getAspectRatio() {
-		return width / height;
+		return width / (float) height;
 	}
 
 	@Override
